+++ truck-js/src/builder.rs
@@ -31,7 +31,7 @@ pub fn circle_arc(vertex0: &Vertex, vertex1: &Vertex, transit: &[f64]) -> Edge {
 #[wasm_bindgen]
 pub fn bezier(vertex0: &Vertex, vertex1: &Vertex, inter_points: &[f64]) -> Edge {
     assert!(
-        inter_points.len() % 3 == 0,
+        inter_points.len().is_multiple_of(3),
         "inter_points cannot convert to 3-dimensional points!"
     );
     let inter_points = inter_points
@@ -139,7 +139,13 @@ pub fn tsweep(shape: &AbstractShape, vector: &[f64]) -> AbstractShape {
 
 /// Sweeps a vertex, an edge, a wire, a face, or a shell by the rotation.
 #[wasm_bindgen]
-pub fn rsweep(shape: &AbstractShape, origin: &[f64], axis: &[f64], angle: f64) -> AbstractShape {
+pub fn rsweep(
+    shape: &AbstractShape,
+    origin: &[f64],
+    axis: &[f64],
+    angle: f64,
+    division: usize,
+) -> AbstractShape {
     intopt!(Point3, origin, Vector3, axis);
-    derive_all_sweepable!(shape, builder::rsweep, (origin, axis, Rad(angle)))
+    derive_all_sweepable!(shape, builder::rsweep, (origin, axis, Rad(angle), division))
 }