+++ truck-modeling/examples/cylinder.rs
@@ -6,7 +6,7 @@ use truck_modeling::*;
 
 fn cylinder(height: f64, radius: f64) -> Solid {
     let vertex = builder::vertex(Point3::new(0.0, -height / 2.0, radius));
-    let circle = builder::rsweep(&vertex, Point3::origin(), Vector3::unit_y(), Rad(7.0));
+    let circle = builder::rsweep(&vertex, Point3::origin(), Vector3::unit_y(), Rad(7.0), 2);
     let disk = builder::try_attach_plane(&[circle]).unwrap();
 
     builder::tsweep(&disk, Vector3::new(0.0, height, 0.0))