+++ truck-platform/tests/common.rs
@@ -141,12 +141,13 @@ pub fn init_device(backends: Backends) -> DeviceHandler {
                 required_features: Default::default(),
                 required_limits: Default::default(),
                 memory_hints: Default::default(),
-                label: None,
+                experimental_features: ExperimentalFeatures::disabled(),
                 trace: Default::default(),
+                label: None,
             })
             .await
             .unwrap();
-        DeviceHandler::new(Arc::new(adapter), Arc::new(device), Arc::new(queue))
+        DeviceHandler::new(adapter, device, queue)
     })
 }