+++ truck-rendimpl/tests/common.rs
@@ -171,12 +171,13 @@ pub fn init_device(instance: &Instance) -> DeviceHandler {
                 required_features: Default::default(),
                 required_limits: Limits::default(),
                 memory_hints: Default::default(),
-                label: None,
                 trace: Default::default(),
+                experimental_features: ExperimentalFeatures::disabled(),
+                label: None,
             })
             .await
             .unwrap();
-        DeviceHandler::new(Arc::new(adapter), Arc::new(device), Arc::new(queue))
+        DeviceHandler::new(adapter, device, queue)
     })
 }