+++ truck-stepio/examples/step-to-mesh.rs
@@ -2,9 +2,12 @@
 
 use clap::Parser;
 use std::path::Path;
+use truck_assembly::assy::*;
 use truck_meshalgo::prelude::*;
-use truck_stepio::r#in::*;
+use truck_stepio::r#in::{convert::*, *};
 use truck_topology::compress::*;
+
+type MeshedCShell = CompressedShell<Point3, PolylineCurve<Point3>, Option<PolygonMesh>>;
 
 #[derive(Parser, Debug)]
 struct Args {
@@ -21,10 +24,6 @@ struct Args {
     /// Ignored when outputting vtk files
     #[arg(long("condition-check"))]
     condition_check: bool,
-    /// output the STEP parse result by json
-    /// When the string "output" is entered, the registered `#i` shell is output "output-i.json".
-    #[arg(long("shape-json"))]
-    shape_json: Option<String>,
 }
 
 fn main() {
@@ -33,29 +32,13 @@ fn main() {
         output_mesh_file,
         only_edge,
         condition_check,
-        shape_json,
     } = Args::parse();
 
     println!("reading file...");
     let step_file = std::fs::read_to_string(input_step_file).unwrap();
-    let exchange = ruststep::parser::parse(&step_file).unwrap();
-    let table = Table::from_data_section(&exchange.data[0]);
+    let table = Table::from_step(&step_file).unwrap();
     println!("meshing...");
-    let polyshells = table
-        .shell
-        .iter()
-        .map(|(idx, shell)| {
-            let shell = table.to_compressed_shell(shell).unwrap();
-            if let Some(output) = &shape_json {
-                let content = serde_json::to_string_pretty(&shell).unwrap();
-                let file_name = format!("{output}-{idx}.json");
-                std::fs::write(file_name, content).unwrap();
-            }
-            let pre = shell.robust_triangulation(0.01).to_polygon();
-            let bdd = pre.bounding_box();
-            shell.robust_triangulation(bdd.diameter() * 0.001)
-        })
-        .collect::<Vec<_>>();
+    let polyshells = step_to_mesh(&table);
 
     let path: &Path = output_mesh_file.as_ref();
     let extension = path.extension().and_then(|e| e.to_str());
@@ -69,11 +52,67 @@ fn main() {
     }
 }
 
-fn output_obj(
-    polyshells: &[CompressedShell<Point3, PolylineCurve<Point3>, Option<PolygonMesh>>],
-    path: &Path,
-    condition_check: bool,
-) {
+fn step_to_mesh<'a>(table: &Table) -> Vec<MeshedCShell> {
+    let assy = table.step_assy().unwrap();
+
+    let node_map = |ProductEntity { shape, attrs }: &ProductEntity| {
+        let shape = shape
+            .iter()
+            .filter_map(|shape| {
+                let shells = match shape {
+                    ProductShape::Solid(solid) => &solid.boundaries,
+                    ProductShape::Shells(shells) => shells,
+                    _ => return None,
+                };
+                let meshed_shells = shells
+                    .into_iter()
+                    .map(|shell| {
+                        let pre = shell.robust_triangulation(0.01).to_polygon();
+                        let bdd = pre.bounding_box();
+                        shell.robust_triangulation(bdd.diameter() * 0.001)
+                    })
+                    .collect::<Vec<_>>();
+                Some(meshed_shells)
+            })
+            .collect::<Vec<_>>();
+        NodeEntity {
+            shape,
+            attrs: attrs.clone(),
+        }
+    };
+    let edge_map = |EdgeEntity { matrix: trans, .. }: &AssembleEntity| EdgeEntity {
+        matrix: Matrix4::try_from(trans).unwrap(),
+        attrs: (),
+    };
+
+    let meshed_assy = assy.par_map(node_map, edge_map);
+
+    meshed_assy
+        .top_nodes()
+        .into_iter()
+        .flat_map(|top| meshed_assy.paths_iter(top.index()))
+        .flat_map(|path| {
+            let matrix = path.matrix();
+            let shapes = path.terminal_node().shape().clone();
+            shapes.into_iter().flatten().map(move |mut shell| {
+                shell.vertices.iter_mut().for_each(|v| {
+                    *v = matrix.transform_point(*v);
+                });
+                shell.edges.iter_mut().for_each(|edge| {
+                    edge.curve.transform_by(matrix);
+                });
+                shell.faces.iter_mut().for_each(|face| {
+                    face.surface
+                        .as_mut()
+                        .map(|surface| surface.transform_by(matrix));
+                });
+                shell
+            })
+        })
+        .collect()
+}
+
+fn output_obj(polyshells: &[MeshedCShell], path: &Path, condition_check: bool) {
     let mut polymesh = PolygonMesh::default();
     polyshells.iter().for_each(|shell| {
         let mut poly = shell.to_polygon();
@@ -85,15 +124,11 @@ fn output_obj(
         }
         polymesh.merge(poly);
     });
-    println!("output obj...");
     let obj_file = std::fs::File::create(path).unwrap();
     obj::write(&polymesh, obj_file).unwrap();
 }
 
-fn output_vtk(
-    polyshells: Vec<CompressedShell<Point3, PolylineCurve<Point3>, Option<PolygonMesh>>>,
-    path: &Path,
-) {
+fn output_vtk(polyshells: Vec<MeshedCShell>, path: &Path) {
     use vtkio::model::*;
     let pieces = polyshells
         .into_iter()
@@ -133,14 +168,11 @@ fn output_vtk(
         data: DataSet::UnstructuredGrid { meta: None, pieces },
     };
 
-    let obj_file = std::fs::File::create(path).unwrap();
-    vtk.write_xml(obj_file).unwrap();
+    let vtk_file = std::fs::File::create(path).unwrap();
+    vtk.write_xml(vtk_file).unwrap();
 }
 
-fn output_vtk_edge(
-    polyshells: Vec<CompressedShell<Point3, PolylineCurve<Point3>, Option<PolygonMesh>>>,
-    path: &Path,
-) {
+fn output_vtk_edge(polyshells: Vec<MeshedCShell>, path: &Path) {
     use vtkio::model::*;
     let pieces = polyshells
         .into_iter()
@@ -175,6 +207,6 @@ fn output_vtk_edge(
         data: DataSet::UnstructuredGrid { meta: None, pieces },
     };
 
-    let obj_file = std::fs::File::create(path).unwrap();
-    vtk.write_xml(obj_file).unwrap();
+    let vtk_file = std::fs::File::create(path).unwrap();
+    vtk.write_xml(vtk_file).unwrap();
 }