+++ truck-stepio/tests/input/main.rs
@@ -1,3 +1,4 @@
+mod assy;
 mod geometry;
 mod table;
 mod tessellate_shape;